Fix branch error
I created a new branch called createNewCardUI, and while working on it, I added a new function in walletUI.js where the system first attempts to load the wallet and bank account from the local system. If this fails, it creates a new bank account and wallet. I did this on a separate branch and tried to merge the changes into the main branch, but something went wrong, and most of the branches in main were deleted and couldn't be retrieved. As a result, I created a new branch from createNewCardUI, which had the latest changes.

**Changes made**

I modified Wallet.js so that loadWallet works by setting the bankAccount parameter to null. This was necessary because when Wallet.CreateWallet is used, a bankAccount class is passed into it, and the _linkBankAccountToWallet() method is automatically called from the constructor to verify the passed bankAccount parameter. However, when Wallet.loadWallet is called, a bankAccount class is not passed in; instead, a string representing the account number is passed to Wallet.loadWallet. It then uses this string to find the wallet, and if the wallet details are found, it builds the wallet class using that data. However, the bankAccount class is not passed into this._bankAccount, and when the constructor automatically calls it, it results in an error.

To rectify this, I added an if statement in _linkBankAccountToWallet() to check that this._bankAccount isn't null before processing. In addition to that, loadWallet now accepts two parameters: sortCode and accountWallet, which are used to retrieve the BankAccount details. These details are then stored inside the wallet as a bank object, resulting in the correct wallet structure.